// examples/multiple-brokers/src/first.broker/moleculer.d.ts
  type GenericObject = { [name: string]: any };

  type LogLevels = 'fatal' | 'error' | 'warn' | 'info' | 'debug' | 'trace';
// ...
  interface LoggerBindings {
    nodeID: string;
    ns: string;
    mod: string;
    svc: string;
    ver: string | void;
  }

  class LoggerInstance {
    fatal(...args: any[]): void;
    error(...args: any[]): void;
    warn(...args: any[]): void;
    info(...args: any[]): void;
    debug(...args: any[]): void;
    trace(...args: any[]): void;
  }

  type ActionHandler<T = any> = ((ctx: Context) => PromiseLike<T> | T) &
    ThisType<Service>;
// ...
  interface BulkheadOptions {
    enabled?: boolean;
    concurrency?: number;
    maxQueueSize?: number;
  }

  interface ActionCacheOptions {
    ttl?: number;
    keys?: Array<string>;
  }

  type ActionVisibility = 'published' | 'public' | 'protected' | 'private';

  interface Action {
    name?: string;
    visibility?: ActionVisibility;
    params?: ActionParams;
    service?: Service;
    cache?: boolean | ActionCacheOptions;
    handler?: ActionHandler;
    metrics?: MetricsOptions;
    bulkhead?: BulkheadOptions;
    circuitBreaker?: BrokerCircuitBreakerOptions;
    retryPolicy?: RetryPolicyOptions;
    fallback?: string | FallbackHandler;

    [key: string]: any;
  }
// ...
  interface ServiceEvent {
    name?: string;
    group?: string;
    handler?: ServiceEventHandler;
  }

  type ServiceEvents = { [key: string]: ServiceEventHandler | ServiceEvent };

  type ServiceMethods = { [key: string]: (...args: any[]) => any } & ThisType<
    Service
  >;

  type Middleware = {
    [name: string]:
      | ((handler: ActionHandler, action: Action) => any)
      | ((handler: ActionHandler, event: ServiceEvent) => any)
      | ((handler: ActionHandler) => any)
      | ((service: Service) => any)
      | ((broker: ServiceBroker) => any);
  };
// ...
  type CheckRetryable = (err: Error) => boolean;

  interface BrokerCircuitBreakerOptions {
    enabled?: boolean;
    threshold?: number;
    windowTime?: number;
    minRequestCount?: number;
    halfOpenTime?: number;
    check?: CheckRetryable;
  }

  interface RetryPolicyOptions {
    enabled?: boolean;
    retries?: number;
    delay?: number;
    maxDelay?: number;
    factor?: number;
    check: CheckRetryable;
  }

  interface BrokerRegistryOptions {
    strategy?: Function | string;
    strategyOptions?: GenericObject;
    preferLocal?: boolean;
  }

  interface BrokerTransitOptions {
    maxQueueSize?: number;
    packetLogFilter?: Array<string>;
    disableReconnect?: boolean;
  }

  interface BrokerTrackingOptions {
    enabled?: boolean;
    shutdownTimeout?: number;
  }

  interface LogLevelConfig {
    [module: string]: boolean | LogLevels;
  }

  interface BrokerOptions {
    namespace?: string;
    nodeID?: string;

    logger?:
      | ((bindings: LoggerBindings) => LoggerInstance)
      | LoggerInstance
      | boolean;
    logLevel?: LogLevels | LogLevelConfig;
    logFormatter?: Function | string;
    logObjectPrinter?: Function;

    transporter?: Transporter | string | GenericObject;
    requestTimeout?: number;
    retryPolicy?: RetryPolicyOptions;

    maxCallLevel?: number;
    heartbeatInterval?: number;
    heartbeatTimeout?: number;

    trackContext?: boolean;
    gracefulStopTimeout?: number;

    tracking?: BrokerTrackingOptions;

    disableBalancer?: boolean;

    registry?: BrokerRegistryOptions;

    circuitBreaker?: BrokerCircuitBreakerOptions;

    bulkhead?: BulkheadOptions;

    transit?: BrokerTransitOptions;

    cacher?: Cacher | string | GenericObject;
    serializer?: Serializer | string | GenericObject;

    validation?: boolean;
    validator?: Validator;
    metrics?: boolean;
    metricsRate?: number;
    internalServices?: boolean;
    internalMiddlewares?: boolean;

    hotReload?: boolean;

    middlewares?: Array<Middleware>;
    replCommands?: Array<GenericObject>;

    ServiceFactory?: Service;
    ContextFactory?: Context;

    created?: (broker: ServiceBroker) => void;
    started?: (broker: ServiceBroker) => void;
    stopped?: (broker: ServiceBroker) => void;

    /**
     * If true, process.on("beforeExit/exit/SIGINT/SIGTERM", ...) handler won't be registered!
     * You have to register this manually and stop broker in this case!
     */
    skipProcessEventRegistration?: boolean;
  }
// ...
  interface PongResponse {
    nodeID: string;
    elapsedTime: number;
    timeDiff: number;
  }

  interface ServiceBroker {
    [name: string]: PongResponse;
  }